+++ tests/integration/models/reddit/test_live.py
@@ -3,7 +3,7 @@ import mock
 import pytest
 
 from praw.const import API_PATH
-from praw.exceptions import APIException
+from praw.exceptions import RedditAPIException
 from praw.models import (
     LiveThread,
     LiveUpdate,
@@ -99,9 +99,12 @@ class TestLiveContributorRelationship(IntegrationTest):
             "TestLiveContributorRelationship_test_invite__already_invited"
         ):
             thread.contributor.invite("nmtake")
-            with pytest.raises(APIException) as excinfo:
+            with pytest.raises(RedditAPIException) as excinfo:
                 thread.contributor.invite("nmtake")
-        assert excinfo.value.error_type == "LIVEUPDATE_ALREADY_CONTRIBUTOR"
+        assert (
+            excinfo.value.items[0].error_type
+            == "LIVEUPDATE_ALREADY_CONTRIBUTOR"
+        )
 
     def test_invite__empty_list(self):
         self.reddit.read_only = False