+++ tests/integration/models/reddit/test_subreddit.py
@@ -18,9 +18,9 @@ from prawcore import (
 
 from praw.const import PNG_HEADER
 from praw.exceptions import (
-    APIException,
     ClientException,
     InvalidFlairTemplateID,
+    RedditAPIException,
     TooLargeMediaException,
     WebSocketException,
 )
@@ -118,7 +118,7 @@ class TestSubreddit(IntegrationTest):
     def test_create__exists(self):
         self.reddit.read_only = False
         with self.recorder.use_cassette("TestSubreddit.test_create__exists"):
-            with pytest.raises(APIException) as excinfo:
+            with pytest.raises(RedditAPIException) as excinfo:
                 self.reddit.subreddit.create(
                     "redditdev",
                     title="redditdev",
@@ -126,14 +126,14 @@ class TestSubreddit(IntegrationTest):
                     subreddit_type="public",
                     wikimode="disabled",
                 )
-            assert excinfo.value.error_type == "SUBREDDIT_EXISTS"
+            assert excinfo.value.items[0].error_type == "SUBREDDIT_EXISTS"
 
     def test_create__invalid_parameter(self):
         self.reddit.read_only = False
         with self.recorder.use_cassette(
             "TestSubreddit.test_create__invalid_parameter"
         ):
-            with pytest.raises(APIException) as excinfo:
+            with pytest.raises(RedditAPIException) as excinfo:
                 # Supplying invalid setting for link_type
                 self.reddit.subreddit.create(
                     name="PRAW_iavynavffv",
@@ -142,14 +142,14 @@ class TestSubreddit(IntegrationTest):
                     subreddit_type="public",
                     wikimode="disabled",
                 )
-            assert excinfo.value.error_type == "INVALID_OPTION"
+            assert excinfo.value.items[0].error_type == "INVALID_OPTION"
 
     def test_create__missing_parameter(self):
         self.reddit.read_only = False
         with self.recorder.use_cassette(
             "TestSubreddit.test_create__missing_parameter"
         ):
-            with pytest.raises(APIException) as excinfo:
+            with pytest.raises(RedditAPIException) as excinfo:
                 # Not supplying required field title.
                 self.reddit.subreddit.create(
                     name="PRAW_iavynavffv",
@@ -158,7 +158,7 @@ class TestSubreddit(IntegrationTest):
                     subreddit_type="public",
                     wikimode="disabled",
                 )
-            assert excinfo.value.error_type == "NO_TEXT"
+            assert excinfo.value.items[0].error_type == "NO_TEXT"
 
     @mock.patch("time.sleep", return_value=None)
     def test_message(self, _):
@@ -341,7 +341,7 @@ class TestSubreddit(IntegrationTest):
                 pytest.placeholders.test_subreddit
             )
             with pytest.raises(
-                (APIException, BadRequest)
+                (RedditAPIException, BadRequest)
             ):  # waiting for prawcore fix
                 subreddit.submit(
                     "dfgfdgfdgdf", url="https://www.google.com",
@@ -473,7 +473,7 @@ class TestSubreddit(IntegrationTest):
             )
             image = self.image_path("test.jpg")
             with pytest.raises(
-                (APIException, BadRequest)
+                (RedditAPIException, BadRequest)
             ):  # waiting for prawcore fix
                 subreddit.submit_image(
                     "gdfgfdgdgdgfgfdgdfgfdgfdg", image,
@@ -672,7 +672,7 @@ class TestSubreddit(IntegrationTest):
             )
             image = self.image_path("test.mov")
             with pytest.raises(
-                (APIException, BadRequest)
+                (RedditAPIException, BadRequest)
             ):  # waiting for prawcore fix
                 subreddit.submit_video("gdfgfdgdgdgfgfdgdfgfdgfdg", image)
 
@@ -1359,9 +1359,9 @@ class TestSubredditModeration(IntegrationTest):
         with self.recorder.use_cassette(
             "TestSubredditModeration.test_accept_invite__no_invite"
         ):
-            with pytest.raises(APIException) as excinfo:
+            with pytest.raises(RedditAPIException) as excinfo:
                 self.subreddit.mod.accept_invite()
-            assert excinfo.value.error_type == "NO_INVITE_FOUND"
+            assert excinfo.value.items[0].error_type == "NO_INVITE_FOUND"
 
     def test_edited(self):
         self.reddit.read_only = False
@@ -1787,11 +1787,11 @@ class TestSubredditRelationships(IntegrationTest):
         with self.recorder.use_cassette(
             "TestSubredditRelationships.moderator_invite__invalid_perm"
         ):
-            with pytest.raises(APIException) as excinfo:
+            with pytest.raises(RedditAPIException) as excinfo:
                 self.subreddit.moderator.invite(
                     self.REDDITOR, permissions=["a"]
                 )
-            assert excinfo.value.error_type == "INVALID_PERMISSIONS"
+            assert excinfo.value.items[0].error_type == "INVALID_PERMISSIONS"
 
     @mock.patch("time.sleep", return_value=None)
     def test_moderator_invite__no_perms(self, _):
@@ -2127,22 +2127,22 @@ class TestSubredditStylesheet(IntegrationTest):
         with self.recorder.use_cassette(
             "TestSubredditStylesheet.test_upload__invalid"
         ):
-            with pytest.raises(APIException) as excinfo:
+            with pytest.raises(RedditAPIException) as excinfo:
                 self.subreddit.stylesheet.upload(
                     "praw", self.image_path("invalid.jpg")
                 )
-        assert excinfo.value.error_type == "IMAGE_ERROR"
+        assert excinfo.value.items[0].error_type == "IMAGE_ERROR"
 
     def test_upload__invalid_ext(self):
         self.reddit.read_only = False
         with self.recorder.use_cassette(
             "TestSubredditStylesheet.test_upload__invalid_ext"
         ):
-            with pytest.raises(APIException) as excinfo:
+            with pytest.raises(RedditAPIException) as excinfo:
                 self.subreddit.stylesheet.upload(
                     "praw.png", self.image_path("white-square.png")
                 )
-        assert excinfo.value.error_type == "BAD_CSS_NAME"
+        assert excinfo.value.items[0].error_type == "BAD_CSS_NAME"
 
     def test_upload__too_large(self):
         self.reddit.read_only = False
@@ -2283,11 +2283,11 @@ class TestSubredditStylesheet(IntegrationTest):
                 "upload_mobile_header",
                 "upload_mobile_icon",
             ]:
-                with pytest.raises(APIException) as excinfo:
+                with pytest.raises(RedditAPIException) as excinfo:
                     getattr(self.subreddit.stylesheet, method)(
                         self.image_path("invalid.jpg")
                     )
-                assert excinfo.value.error_type == "IMAGE_ERROR"
+                assert excinfo.value.items[0].error_type == "IMAGE_ERROR"
 
     def test_upload__others_too_large(self):
         self.reddit.read_only = False