+++ tests/integration/test_objector.py
@@ -1,7 +1,7 @@
 """Test praw.objector."""
 import pytest
 
-from praw.exceptions import APIException
+from praw.exceptions import RedditAPIException
 
 from . import IntegrationTest
 
@@ -13,7 +13,7 @@ class TestObjector(IntegrationTest):
             "TestObjector.test_raise_api_exception"
         ):
             submission = self.reddit.submission("4b536h")
-            with pytest.raises(APIException) as excinfo:
+            with pytest.raises(RedditAPIException) as excinfo:
                 submission.mod.approve()
-            assert excinfo.value.error_type == "USER_REQUIRED"
-            assert str(excinfo.value) == message
+            assert excinfo.value.items[0].error_type == "USER_REQUIRED"
+            assert str(excinfo.value.items[0]) == message